requirements.txt
pip~=20.0.2
keyring~=18.0.1
setuptools~=45.2.0
Django~=3.1.4
cryptography~=3.4.6
pytz~=2019.3
wheel~=0.34.2
sqlparse~=0.4.1
mysqlclient~=2.0.3
asgiref~=3.3.1
colorama~=0.4.3
PyYAML~=5.3.1
numpy~=1.19.4
djangorestframework~=3.12.2
MarkupSafe~=1.1.0
httplib2~=0.14.0
certifi~=2020.12.5
chardet~=3.0.4
requests~=2.22.0
Click~=7.0
urllib3~=1.25.11
six~=1.15.0
oauthlib~=3.1.0
blinker~=1.4
idna~=2.8
simplejson~=3.16.0
pycparser~=2.20
lxml~=4.6.1
PyJWT~=1.7.1
bcrypt~=3.1.7